Don't expose context and check existence of credentials functions
