+++ src/@sdk/react/helpers.ts
@@ -1,6 +1,20 @@
+import ApolloClient from "apollo-client";
 import React from "react";
 
 import { getAuthToken, removeAuthToken } from "../auth";
+import { ApolloContext } from "./context";
+
+export function useSaleorClient<TCache = object>(): ApolloClient<TCache> {
+  const client = React.useContext(ApolloContext);
+
+  if (!client) {
+    throw new Error(
+      "Could not find Saleor's apollo client in the context. " +
+        "Did you forget to wrap the root component in a <SaleorProvider>?"
+    );
+  }
+  return client;
+}
 
 export const useAuth = (
   stateChangeCallback?: (authenticated?: boolean) => void
