+++ src/@sdk/react/index.ts
@@ -1,5 +1,4 @@
 export * from "./components";
-export * from "./context";
 export * from "./queries";
 export * from "./mutations";
 export * from "./helpers";
