+++ src/@sdk/react/mutations.ts
@@ -17,13 +17,15 @@ export const useSignIn = (
     update: (proxy, data) => {
       if (data.data.tokenCreate.token) {
         setAuthToken(data.data.tokenCreate.token);
-        if (window.PasswordCredential && variables) {
-          navigator.credentials.store(
-            new window.PasswordCredential({
-              id: variables.email,
-              password: variables.password,
-            })
-          );
+        if (navigator.credentials && navigator.credentials.store) {
+          if (window.PasswordCredential && variables) {
+            navigator.credentials.store(
+              new window.PasswordCredential({
+                id: variables.email,
+                password: variables.password,
+              })
+            );
+          }
         }
       }
       if (options && options.update) {
